requirements.txt
fastapi[standard]
aio-pika
redis
python-dotenv
sqlalchemy
psycopg2